Update Goldman Sachs funds (formerly NN funds)
NN was acquired by Goldman Sachs. This change updates the asset manager
logo, the fund names and the fund symbols as the NN name is no longer
used by Goldman Sachs.

This change also corrects the KIID download link and updates the TER and
internal transaction costs as per costs shown for these funds on:
https://www.gsam.com/responsible-investing/nl-NL/professional/funds